requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.2.2
mysql-connector-python==8.2.0